+++ nd4j/nd4j-backends/nd4j-api-parent/nd4j-api/src/main/java/org/nd4j/autodiff/samediff/ops/SDNN.java
@@ -446,7 +446,7 @@ public class SDNN extends SDOps {
    * @return output Output variable (NUMERIC type)
    */
   public SDVariable layerNorm(SDVariable input, SDVariable gain, SDVariable bias,
-      boolean channelsFirst, int... dimensions) {
+      boolean channelsFirst, long... dimensions) {
     SDValidation.validateNumerical("layerNorm", "input", input);
     SDValidation.validateNumerical("layerNorm", "gain", gain);
     SDValidation.validateNumerical("layerNorm", "bias", bias);
@@ -468,7 +468,7 @@ public class SDNN extends SDOps {
    * @return output Output variable (NUMERIC type)
    */
   public SDVariable layerNorm(String name, SDVariable input, SDVariable gain, SDVariable bias,
-      boolean channelsFirst, int... dimensions) {
+      boolean channelsFirst, long... dimensions) {
     SDValidation.validateNumerical("layerNorm", "input", input);
     SDValidation.validateNumerical("layerNorm", "gain", gain);
     SDValidation.validateNumerical("layerNorm", "bias", bias);
@@ -489,7 +489,7 @@ public class SDNN extends SDOps {
    * @return output Output variable (NUMERIC type)
    */
   public SDVariable layerNorm(SDVariable input, SDVariable gain, boolean channelsFirst,
-      int... dimensions) {
+      long... dimensions) {
     SDValidation.validateNumerical("layerNorm", "input", input);
     SDValidation.validateNumerical("layerNorm", "gain", gain);
     Preconditions.checkArgument(dimensions.length >= 1, "dimensions has incorrect size/length. Expected: dimensions.length >= 1, got %s", dimensions.length);
@@ -509,7 +509,7 @@ public class SDNN extends SDOps {
    * @return output Output variable (NUMERIC type)
    */
   public SDVariable layerNorm(String name, SDVariable input, SDVariable gain, boolean channelsFirst,
-      int... dimensions) {
+      long... dimensions) {
     SDValidation.validateNumerical("layerNorm", "input", input);
     SDValidation.validateNumerical("layerNorm", "gain", gain);
     Preconditions.checkArgument(dimensions.length >= 1, "dimensions has incorrect size/length. Expected: dimensions.length >= 1, got %s", dimensions.length);
@@ -1161,7 +1161,6 @@ public class SDNN extends SDOps {
     SDVariable out =  new org.nd4j.linalg.api.ops.impl.transforms.custom.SoftMax(sd,x, -1).outputVariable();
     return sd.updateVariableNameAndReference(out, name);
   }
-
 
   /**
    * Element-wise softplus function: out = log(exp(x) + 1)<br>
